+++ hackage2nix/Main.hs
@@ -157,8 +157,8 @@ main = do
           drv = fromGenericPackageDescription haskellResolver nixpkgsResolver targetPlatform (compilerInfo config) flagAssignment [] descr
                   & src .~ urlDerivationSource ("mirror://hackage/" ++ display pkgId ++ ".tar.gz") tarballSHA256
                   & editedCabalFile .~ cabalSHA256
-                  & metaSection.platforms .~ fmap (Set.map NixpkgsPlatformSingle) (Map.lookup name (supportedPlatforms config))
-                  & metaSection.badPlatforms .~ fmap (Set.map NixpkgsPlatformSingle) (Map.lookup name (unsupportedPlatforms config))
+                  & metaSection.platforms .~ Map.lookup name (supportedPlatforms config)
+                  & metaSection.badPlatforms .~ Map.lookup name (unsupportedPlatforms config)
                   & metaSection.hydraPlatforms %~ (if isHydraEnabled then id else const (Just Set.empty))
                   & metaSection.broken ||~ isBroken
                   & metaSection.maintainers .~ Map.findWithDefault Set.empty name globalPackageMaintainers